# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Millicall PBX",
    "version": "0.1.0"
  },
  "paths": {
...
    "/provisioning/Panasonic/Config{mac_address}.cfg": {
      "get": {
        "tags": [
          "provisioning"
        ],
        "summary": "Device Config",
        "operationId": "device_config_provisioning_Panasonic_Config_mac_address__cfg_get",
        "parameters": [
          {
            "name": "mac_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mac Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/provisioning/Yealink/{boot_file}.boot": {
      "get": {
        "tags": [
          "provisioning"
        ],
        "summary": "Yealink Model Boot",
        "description": "Model-specific boot file (e.g. y000000000042 for T43U) — same content.",
        "operationId": "yealink_model_boot_provisioning_Yealink__boot_file__boot_get",
        "parameters": [
          {
            "name": "boot_file",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Boot File"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/provisioning/Yealink/{mac_address}.cfg": {
      "get": {
        "tags": [
          "provisioning"
        ],
        "summary": "Yealink Device Config",
        "description": "Per-device config — MAC address without separators (e.g. 805ec0a1b2c3.cfg).",
        "operationId": "yealink_device_config_provisioning_Yealink__mac_address__cfg_get",
        "parameters": [
          {
            "name": "mac_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mac Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}